+++ core/src/rules/mod.rs
@@ -4,6 +4,7 @@ pub mod eof_newline;
 pub mod formatting_toggle;
 pub mod generics_consolidator;
 pub mod ignore_asm_instructions;
+pub mod lowercase_keywords;
 pub mod optimising_line_formatter;
 pub mod token_spacing;
 
@@ -13,5 +14,6 @@ pub use eof_newline::*;
 pub use formatting_toggle::*;
 pub use generics_consolidator::*;
 pub use ignore_asm_instructions::*;
+pub use lowercase_keywords::*;
 pub use optimising_line_formatter::*;
 pub use token_spacing::*;